alchemy_backend_fastapi/requirements.txt
fastapi>=0.122.0
uvicorn[standard]>=0.32.0
supabase>=2.10.0
python-dotenv>=1.0.0
requests>=2.32.0
python-multipart>=0.0.12
pyjwt>=2.9.0
bcrypt>=4.2.0
psycopg2-binary>=2.9.10
pydantic>=2.12.5
starlette>=0.50.0
